+++ src/wl_ngram_generator.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Wordless: N-gram
+# Wordless: N-gram Generator
 # Copyright (C) 2018-2022  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify
@@ -34,11 +34,11 @@ from wl_nlp import wl_matching, wl_nlp_utils, wl_texts, wl_token_processing
 from wl_utils import wl_misc, wl_msgs, wl_sorting, wl_threading
 from wl_widgets import wl_boxes, wl_layouts, wl_tables, wl_widgets
 
-class Wl_Table_Ngram(wl_tables.Wl_Table_Data_Filter_Search):
+class Wl_Table_Ngram_Generator(wl_tables.Wl_Table_Data_Filter_Search):
     def __init__(self, parent):
         super().__init__(
             parent,
-            tab = 'ngram',
+            tab = 'ngram_generator',
             headers = [
                 parent.tr('Rank'),
                 parent.tr('N-gram'),
@@ -61,27 +61,27 @@ class Wl_Table_Ngram(wl_tables.Wl_Table_Data_Filter_Search):
         self.button_generate_table.clicked.connect(lambda: generate_table(self.main, self))
         self.button_generate_fig.clicked.connect(lambda: generate_fig(self.main))
 
-class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
+class Wrapper_Ngram_Generator(wl_layouts.Wl_Wrapper):
     def __init__(self, main):
         super().__init__(main)
 
         # Table
-        self.table_ngram = Wl_Table_Ngram(self)
+        self.table_ngram_generator = Wl_Table_Ngram_Generator(self)
 
         layout_results = wl_layouts.Wl_Layout()
-        layout_results.addWidget(self.table_ngram.label_number_results, 0, 0)
-        layout_results.addWidget(self.table_ngram.button_results_filter, 0, 2)
-        layout_results.addWidget(self.table_ngram.button_results_search, 0, 3)
+        layout_results.addWidget(self.table_ngram_generator.label_number_results, 0, 0)
+        layout_results.addWidget(self.table_ngram_generator.button_results_filter, 0, 2)
+        layout_results.addWidget(self.table_ngram_generator.button_results_search, 0, 3)
 
         layout_results.setColumnStretch(1, 1)
 
         self.wrapper_table.layout().addLayout(layout_results, 0, 0, 1, 5)
-        self.wrapper_table.layout().addWidget(self.table_ngram, 1, 0, 1, 5)
-        self.wrapper_table.layout().addWidget(self.table_ngram.button_generate_table, 2, 0)
-        self.wrapper_table.layout().addWidget(self.table_ngram.button_generate_fig, 2, 1)
-        self.wrapper_table.layout().addWidget(self.table_ngram.button_exp_selected, 2, 2)
-        self.wrapper_table.layout().addWidget(self.table_ngram.button_exp_all, 2, 3)
-        self.wrapper_table.layout().addWidget(self.table_ngram.button_clr, 2, 4)
+        self.wrapper_table.layout().addWidget(self.table_ngram_generator, 1, 0, 1, 5)
+        self.wrapper_table.layout().addWidget(self.table_ngram_generator.button_generate_table, 2, 0)
+        self.wrapper_table.layout().addWidget(self.table_ngram_generator.button_generate_fig, 2, 1)
+        self.wrapper_table.layout().addWidget(self.table_ngram_generator.button_exp_selected, 2, 2)
+        self.wrapper_table.layout().addWidget(self.table_ngram_generator.button_exp_all, 2, 3)
+        self.wrapper_table.layout().addWidget(self.table_ngram_generator.button_clr, 2, 4)
 
         # Token Settings
         self.group_box_token_settings = QGroupBox(self.tr('Token Settings'), self)
@@ -157,7 +157,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
             self.checkbox_match_tags
         ) = wl_widgets.wl_widgets_search_settings(
             main,
-            tab = 'ngram'
+            tab = 'ngram_generator'
         )
 
         self.label_search_term_position = QLabel(self.tr('Search Term Position:'), self)
@@ -180,7 +180,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
             self.button_context_settings
         ) = wl_widgets.wl_widgets_context_settings(
             self,
-            tab = 'ngram'
+            tab = 'ngram_generator'
         )
 
         self.group_box_search_settings.setCheckable(True)
@@ -189,7 +189,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
 
         self.checkbox_multi_search_mode.stateChanged.connect(self.search_settings_changed)
         self.line_edit_search_term.textChanged.connect(self.search_settings_changed)
-        self.line_edit_search_term.returnPressed.connect(self.table_ngram.button_generate_table.click)
+        self.line_edit_search_term.returnPressed.connect(self.table_ngram_generator.button_generate_table.click)
         self.list_search_terms.model().dataChanged.connect(self.search_settings_changed)
 
         self.checkbox_ignore_case.stateChanged.connect(self.search_settings_changed)
@@ -332,7 +332,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
             self.checkbox_show_breakdown
         ) = wl_widgets.wl_widgets_table_settings(
             self,
-            tables = [self.table_ngram]
+            tables = [self.table_ngram_generator]
         )
 
         self.checkbox_show_pct.stateChanged.connect(self.table_settings_changed)
@@ -422,9 +422,9 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
 
     def load_settings(self, defaults = False):
         if defaults:
-            settings = copy.deepcopy(self.main.settings_default['ngram'])
+            settings = copy.deepcopy(self.main.settings_default['ngram_generator'])
         else:
-            settings = copy.deepcopy(self.main.settings_custom['ngram'])
+            settings = copy.deepcopy(self.main.settings_custom['ngram_generator'])
 
         # Token Settings
         self.checkbox_words.setChecked(settings['token_settings']['words'])
@@ -466,7 +466,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
 
         # Context Settings
         if defaults:
-            self.main.wl_context_settings_ngram.load_settings(defaults = True)
+            self.main.wl_context_settings_ngram_generator.load_settings(defaults = True)
 
         # Generation Settings
         self.checkbox_ngram_size_sync.setChecked(settings['generation_settings']['ngram_size_sync'])
@@ -502,7 +502,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
         self.fig_settings_changed()
 
     def token_settings_changed(self):
-        settings = self.main.settings_custom['ngram']['token_settings']
+        settings = self.main.settings_custom['ngram_generator']['token_settings']
 
         settings['words'] = self.checkbox_words.isChecked()
         settings['lowercase'] = self.checkbox_lowercase.isChecked()
@@ -528,10 +528,10 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
 
             self.group_box_search_settings.setChecked(False)
 
-        self.main.wl_context_settings_ngram.token_settings_changed()
+        self.main.wl_context_settings_ngram_generator.token_settings_changed()
 
     def search_settings_changed(self):
-        settings = self.main.settings_custom['ngram']['search_settings']
+        settings = self.main.settings_custom['ngram_generator']['search_settings']
 
         settings['search_settings'] = self.group_box_search_settings.isChecked()
 
@@ -554,7 +554,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
         settings['allow_skipped_tokens_within_search_terms'] = self.checkbox_allow_skipped_tokens_within_search_terms.isChecked()
 
     def generation_settings_changed(self):
-        settings = self.main.settings_custom['ngram']['generation_settings']
+        settings = self.main.settings_custom['ngram_generator']['generation_settings']
 
         settings['ngram_size_sync'] = self.checkbox_ngram_size_sync.isChecked()
         settings['ngram_size_min'] = self.spin_box_ngram_size_min.value()
@@ -579,7 +579,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
         if settings['allow_skipped_tokens']:
             self.spin_box_allow_skipped_tokens.setEnabled(True)
 
-            if self.main.settings_custom['ngram']['search_settings']['search_settings']:
+            if self.main.settings_custom['ngram_generator']['search_settings']['search_settings']:
                 self.checkbox_allow_skipped_tokens_within_search_terms.setEnabled(True)
         else:
             self.spin_box_allow_skipped_tokens.setEnabled(False)
@@ -599,17 +599,17 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
         if self.combo_box_use_data.findText(use_data_old) > -1:
             self.combo_box_use_data.setCurrentText(use_data_old)
         else:
-            self.combo_box_use_data.setCurrentText(self.main.settings_default['ngram']['fig_settings']['use_data'])
+            self.combo_box_use_data.setCurrentText(self.main.settings_default['ngram_generator']['fig_settings']['use_data'])
 
     def table_settings_changed(self):
-        settings = self.main.settings_custom['ngram']['table_settings']
+        settings = self.main.settings_custom['ngram_generator']['table_settings']
 
         settings['show_pct'] = self.checkbox_show_pct.isChecked()
         settings['show_cumulative'] = self.checkbox_show_cumulative.isChecked()
         settings['show_breakdown'] = self.checkbox_show_breakdown.isChecked()
 
     def fig_settings_changed(self):
-        settings = self.main.settings_custom['ngram']['fig_settings']
+        settings = self.main.settings_custom['ngram_generator']['fig_settings']
 
         settings['graph_type'] = self.combo_box_graph_type.currentText()
         settings['sort_by_file'] = self.combo_box_sort_by_file.currentText()
@@ -622,7 +622,7 @@ class Wrapper_Ngram(wl_layouts.Wl_Wrapper):
         settings['rank_max'] = self.spin_box_rank_max.value()
         settings['rank_max_no_limit'] = self.checkbox_rank_max_no_limit.isChecked()
 
-class Wl_Worker_Ngram(wl_threading.Wl_Worker):
+class Wl_Worker_Ngram_Generator(wl_threading.Wl_Worker):
     worker_done = pyqtSignal(str, dict, dict, dict)
 
     def __init__(self, main, dialog_progress, update_gui):
@@ -637,7 +637,7 @@ class Wl_Worker_Ngram(wl_threading.Wl_Worker):
         try:
             texts = []
 
-            settings = self.main.settings_custom['ngram']
+            settings = self.main.settings_custom['ngram_generator']
 
             ngram_size_min = settings['generation_settings']['ngram_size_min']
             ngram_size_max = settings['generation_settings']['ngram_size_max']
@@ -651,7 +651,7 @@ class Wl_Worker_Ngram(wl_threading.Wl_Worker):
                 ngrams = []
 
                 text = copy.deepcopy(file['text'])
-                text = wl_token_processing.wl_process_tokens_ngram(
+                text = wl_token_processing.wl_process_tokens_ngram_generator(
                     self.main, text,
                     token_settings = settings['token_settings']
                 )
@@ -874,7 +874,7 @@ class Wl_Worker_Ngram(wl_threading.Wl_Worker):
         except Exception:
             self.err_msg = traceback.format_exc()
 
-class Wl_Worker_Ngram_Table(Wl_Worker_Ngram):
+class Wl_Worker_Ngram_Generator_Table(Wl_Worker_Ngram_Generator):
     def run(self):
         super().run()
 
@@ -886,7 +886,7 @@ class Wl_Worker_Ngram_Table(Wl_Worker_Ngram):
             self.ngrams_text
         )
 
-class Wl_Worker_Ngram_Fig(Wl_Worker_Ngram):
+class Wl_Worker_Ngram_Generator_Fig(Wl_Worker_Ngram_Generator):
     def run(self):
         super().run()
 
@@ -1025,14 +1025,12 @@ def generate_table(main, table):
                 wl_msgs.wl_msg_generate_table_success(main)
             else:
                 wl_msg_boxes.wl_msg_box_no_results(main)
-
                 wl_msgs.wl_msg_generate_table_error(main)
         else:
             wl_dialogs_errs.Wl_Dialog_Err_Fatal(main, err_msg).open()
-
             wl_msgs.wl_msg_fatal_error(main)
 
-    settings = main.settings_custom['ngram']
+    settings = main.settings_custom['ngram_generator']
     files = list(main.wl_file_area.get_selected_files())
 
     if wl_checking_files.check_files_on_loading(main, files):
@@ -1041,19 +1039,14 @@ def generate_table(main, table):
             or not settings['search_settings']['multi_search_mode'] and settings['search_settings']['search_term']
             or settings['search_settings']['multi_search_mode'] and settings['search_settings']['search_terms']
         ):
-            dialog_progress = wl_dialogs_misc.Wl_Dialog_Progress_Process_Data(main)
-
-            worker_ngram_table = Wl_Worker_Ngram_Table(
+            worker_ngram_generator_table = Wl_Worker_Ngram_Generator_Table(
                 main,
-                dialog_progress = dialog_progress,
+                dialog_progress = wl_dialogs_misc.Wl_Dialog_Progress_Process_Data(main),
                 update_gui = update_gui
             )
-
-            thread_ngram_table = wl_threading.Wl_Thread(worker_ngram_table)
-            thread_ngram_table.start_worker()
+            wl_threading.Wl_Thread(worker_ngram_generator_table).start_worker()
         else:
             wl_msg_boxes.wl_msg_box_missing_search_terms_optional(main)
-
             wl_msgs.wl_msg_generate_table_error(main)
     else:
         wl_msgs.wl_msg_generate_table_error(main)
@@ -1111,11 +1104,9 @@ def generate_fig(main):
                 wl_msgs.wl_msg_generate_fig_success(main)
             else:
                 wl_msg_boxes.wl_msg_box_no_results(main)
-
                 wl_msgs.wl_msg_generate_fig_error(main)
         else:
             wl_dialogs_errs.Wl_Dialog_Err_Fatal(main, err_msg).open()
-
             wl_msgs.wl_msg_fatal_error(main)
 
         dialog_progress.accept()
@@ -1123,7 +1114,7 @@ def generate_fig(main):
         if ngrams_freq_files:
             wl_figs.show_fig()
 
-    settings = main.settings_custom['ngram']
+    settings = main.settings_custom['ngram_generator']
     files = list(main.wl_file_area.get_selected_files())
 
     if wl_checking_files.check_files_on_loading(main, files):
@@ -1134,16 +1125,14 @@ def generate_fig(main):
         ):
             dialog_progress = wl_dialogs_misc.Wl_Dialog_Progress_Process_Data(main)
 
-            worker_ngram_fig = Wl_Worker_Ngram_Fig(
+            worker_ngram_generator_fig = Wl_Worker_Ngram_Generator_Fig(
                 main,
                 dialog_progress = dialog_progress,
                 update_gui = update_gui
             )
-            thread_ngram_fig = wl_threading.Wl_Thread(worker_ngram_fig)
-            thread_ngram_fig.start_worker()
+            wl_threading.Wl_Thread(worker_ngram_generator_fig).start_worker()
         else:
             wl_msg_boxes.wl_msg_box_missing_search_terms_optional(main)
-
             wl_msgs.wl_msg_generate_fig_error(main)
     else:
-        wl_msgs.wl_msg_generate_table_error(main)
+        wl_msgs.wl_msg_generate_fig_error(main)